requirements.txt
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-Sock==0.7.0
psycopg2-binary==2.9.10
pytest==8.3.5
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
websocket-client==1.8.0
Werkzeug==3.1.3
